+++ Wabbajack.CLI/Program.cs
@@ -18,6 +18,7 @@ using Wabbajack.Server.Lib;
 using Wabbajack.Services.OSIntegrated;
 using Wabbajack.VFS;
 using Client = Wabbajack.Networking.GitHub.Client;
+using Wabbajack.CLI.Builder;
 
 namespace Wabbajack.CLI;
 
@@ -46,7 +47,13 @@ internal class Program
 
 
                 services.AddTransient<Context>();
-                services.AddCommands();
+                
+                services.AddSingleton<CommandLineBuilder>();
+                services.AddCLIVerbs();
+
+ 
+
+                
                 services.AddSingleton<IUserInterventionHandler, UserInterventionHandler>();
             }).Build();