+++ Wabbajack.CLI/Verbs/Compile.cs
@@ -4,6 +4,7 @@ using System.CommandLine.NamingConventionBinder;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
+using Wabbajack.CLI.Builder;
 using Wabbajack.Compiler;
 using Wabbajack.Downloaders;
 using Wabbajack.Downloaders.GameFile;
@@ -14,7 +15,7 @@ using Wabbajack.VFS;
 
 namespace Wabbajack.CLI.Verbs;
 
-public class Compile : IVerb
+public class Compile
 {
     private readonly ILogger<Compile> _logger;
     private readonly Client _wjClient;