+++ Wabbajack.CLI/Verbs/Decrypt.cs
@@ -4,13 +4,14 @@ using System.CommandLine.Invocation;
 using System.CommandLine.NamingConventionBinder;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
+using Wabbajack.CLI.Builder;
 using Wabbajack.Paths;
 using Wabbajack.Paths.IO;
 using Wabbajack.Services.OSIntegrated;
 
 namespace Wabbajack.CLI.Verbs;
 
-public class Decrypt : IVerb
+public class Decrypt
 {
     private readonly ILogger<Decrypt> _logger;