+++ Wabbajack.CLI/Verbs/DownloadAll.cs
@@ -6,6 +6,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
+using Wabbajack.CLI.Builder;
 using Wabbajack.Common;
 using Wabbajack.Downloaders;
 using Wabbajack.DTOs;
@@ -20,7 +21,7 @@ using Wabbajack.VFS;
 
 namespace Wabbajack.CLI.Verbs;
 
-public class DownloadAll : IVerb
+public class DownloadAll
 {
     private readonly DownloadDispatcher _dispatcher;
     private readonly ILogger<DownloadAll> _logger;