+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### Changelog
 
+#### Version - 3.0.2.0 - 10/14/2022
+* Show Modlist readmes after install
+* Basic support for commandline options in the WPF app
+
 #### Version - 3.0.1.9 - 10/4/2022
 * Lots of compiler improvements for faster compilation
 * Limit the log view to the last 200 messages for better UI performance